requirements.txt
discord.py
google-api-python-client
youtube-dl
requests